requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.9.3
configparser==3.5.0b2
cryptography==2.6.1
cycler==0.10.0
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
gyp==0.1
ipaddress==1.0.17
jdcal==1.4.1
keepnote==0.7.8
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.16.2
olefile==0.46
openpyxl==2.6.4
pandas==0.24.2
Pillow==5.4.1
pycairo==1.16.2
pycrypto==2.6.1
PyGObject==3.30.4
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2019.1
pyusb==1.1.0
pyxdg==0.25
randomcolor==0.4.4.5
SecretStorage==2.3.1
six==1.12.0
soupsieve==1.9.6
subprocess32==3.5.3
vboxapi==1.0
xlwt==1.3.0
